+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is base on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), 
 
 ## [0.56.00.000] - TBD
 ### Added
-- Use Confluent Cloud for Apache Flink to handle Avro Flight Consolidator App.
+- Use Confluent Cloud for Apache Flink to handle the Avro Flight Consolidator App.
+- Use Confluent Cloud for Apache Flink to handle the Avro Flyer Stats App.
 
 ## [0.55.00.000] - 2024-12-16
 ### Added