+++ CHANGELOG.md
@@ -3,10 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is base on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.56.00.000] - TBD
+## [0.56.00.000] - 2024-12-21
 ### Added
 - Use Confluent Cloud for Apache Flink to handle the Avro Flight Consolidator App.
-- Use Confluent Cloud for Apache Flink to handle the Avro Flyer Stats App.
 
 ## [0.55.00.000] - 2024-12-16
 ### Added